app/express-app/requirements_win.txt
asf_search==7.0.9
certifi==2024.2.2
charset-normalizer==3.3.2
dateparser==1.2.0
idna==3.7
importlib_metadata==7.1.0
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.16
requests==2.31.0
shapely==2.0.4
six==1.16.0
tenacity==8.2.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
zipp==3.18.1
pillow==10.3.0
numpy
rioxarray